Listing of packages made as subcommand
