+++ ops.go
@@ -9,14 +9,13 @@ import (
 	"os/exec"
 	"path"
 	"path/filepath"
-	"regexp"
 	"runtime"
 	"strconv"
 	"strings"
 
 	"github.com/go-errors/errors"
+	"github.com/nanovms/ops/cmd"
 	api "github.com/nanovms/ops/lepton"
-	"github.com/olekukonko/tablewriter"
 	"github.com/spf13/cobra"
 )
 
@@ -654,59 +653,6 @@ func InitDefaultRunConfigs(c *api.Config, ports []int) {
 	c.RunConfig.Ports = append(c.RunConfig.Ports, ports...)
 }
 
-func cmdListPackages(cmd *cobra.Command, args []string) {
-	packages := api.GetPackageList()
-
-	searchRegex, err := cmd.Flags().GetString("search")
-	if err != nil {
-		panic(err)
-	}
-
-	table := tablewriter.NewWriter(os.Stdout)
-	table.SetHeader([]string{"PackageName", "Version", "Language", "Runtime", "Description"})
-	table.SetHeaderColor(
-		tablewriter.Colors{tablewriter.Bold, tablewriter.FgWhiteColor},
-		tablewriter.Colors{tablewriter.Bold, tablewriter.FgWhiteColor},
-		tablewriter.Colors{tablewriter.Bold, tablewriter.FgWhiteColor},
-		tablewriter.Colors{tablewriter.Bold, tablewriter.FgWhiteColor},
-		tablewriter.Colors{tablewriter.Bold, tablewriter.FgWhiteColor})
-
-	table.SetRowLine(true)
-
-	var r *regexp.Regexp
-	var filter bool
-	if len(searchRegex) > 0 {
-		filter = true
-		r, err = regexp.Compile(searchRegex)
-		if err != nil {
-			// If the regex cannot compile do not attempt to filter
-			filter = false
-		}
-	}
-
-	for key, val := range *packages {
-		var row []string
-		// If we are told to filter and get no matches then filter out the
-		// current row. If we are not told to filter then just add the
-		// row.
-		if filter &&
-			!(r.MatchString(val.Language) ||
-				r.MatchString(val.Runtime) ||
-				r.MatchString(key)) {
-			continue
-		}
-
-		row = append(row, key)
-		row = append(row, val.Version)
-		row = append(row, val.Language)
-		row = append(row, val.Runtime)
-		row = append(row, val.Description)
-		table.Append(row)
-	}
-
-	table.Render()
-}
-
 func main() {
 
 	var cmdRun = &cobra.Command{
@@ -724,7 +670,7 @@ func main() {
 	var verbose bool
 	var bridged bool
 	var nightly bool
-	var search string
+	//var search string
 	var tap string
 	var targetRoot string
 	var targetCloud string
@@ -794,13 +740,6 @@ func main() {
 		Short: "Version",
 		Run:   printVersion,
 	}
-
-	var cmdList = &cobra.Command{
-		Use:   "list",
-		Short: "list packages",
-		Run:   cmdListPackages,
-	}
-	cmdList.PersistentFlags().StringVarP(&search, "search", "s", "", "search package list")
 
 	var cmdLoadPackage = &cobra.Command{
 		Use:   "load [packagename]",
@@ -878,7 +817,7 @@ func main() {
 	rootCmd.AddCommand(cmdPrintConfig)
 	rootCmd.AddCommand(cmdVersion)
 	rootCmd.AddCommand(cmdUpdate)
-	rootCmd.AddCommand(cmdList)
+	rootCmd.AddCommand(cmd.PackageCommand())
 	rootCmd.AddCommand(cmdLoadPackage)
 	rootCmd.AddCommand(cmdGetPackage)
 	rootCmd.AddCommand(cmdInstance)
